# hc-129210: GC optimisations for case where no objects being collected have finalizers