--- a/Misc/NEWS.d/next/Core_and_Builtins/2025-04-16-12-02-26.gh-issue-129210.a5uLpE.rst
+++ b/Misc/NEWS.d/next/Core_and_Builtins/2025-04-16-12-02-26.gh-issue-129210.a5uLpE.rst
@@ -1,0 +1,2 @@
+Optimise garbage collection in case where objects to collect have
+finalizers, and hence cannot be resurrected.