@duaneg 2025-04-14
Given this is just an optimisation for the incremental GC, which is being introduced in v3.14, I don't imagine it requires a blurb, but I'm happy to add one if it is felt useful.

I've also had a quick look at the free-threading GC and it appears at first glance like similar optimisations are possible for it also. I'm investigating that now and if it pans out will add commits for it to this PR.