@duaneg 2025-04-15
I've just finished running `pyperformance --rigorous --min-time 2` with its default suite, minus a few benchmarks that are failing for unrelated reasons.

<details>
<summary>Out of 102 tests, 20 are "significantly" faster and 7 slower</summary>

```
### 2to3 ###
Mean +- std dev: 197 ms +- 1 ms -> 192 ms +- 1 ms: 1.03x faster
Significant (t=58.77)
--
### async_tree_eager_io_tg ###
Mean +- std dev: 508 ms +- 6 ms -> 497 ms +- 6 ms: 1.02x faster
Significant (t=13.52)
--
### async_tree_eager_tg ###
Mean +- std dev: 175 ms +- 2 ms -> 171 ms +- 1 ms: 1.03x faster
Significant (t=22.28)
--
### async_tree_io_tg ###
Mean +- std dev: 511 ms +- 7 ms -> 501 ms +- 9 ms: 1.02x faster
Significant (t=10.46)
--
### asyncio_tcp ###
Mean +- std dev: 272 ms +- 5 ms -> 266 ms +- 6 ms: 1.02x faster
Significant (t=8.03)
--
### coroutines ###
Mean +- std dev: 16.5 ms +- 0.4 ms -> 17.3 ms +- 0.4 ms: 1.05x slower
Significant (t=-15.60)
--
### coverage ###
Mean +- std dev: 80.1 ms +- 1.3 ms -> 73.6 ms +- 1.3 ms: 1.09x faster
Significant (t=38.51)
--
### create_gc_cycles ###
Mean +- std dev: 1.43 ms +- 0.02 ms -> 1.39 ms +- 0.02 ms: 1.03x faster
Significant (t=22.63)
--
### deltablue ###
Mean +- std dev: 2.33 ms +- 0.05 ms -> 2.28 ms +- 0.05 ms: 1.02x faster
Significant (t=7.98)
--
### docutils ###
Mean +- std dev: 1.96 sec +- 0.01 sec -> 1.91 sec +- 0.01 sec: 1.02x faster
Significant (t=25.21)
--
### html5lib ###
Mean +- std dev: 56.4 ms +- 0.9 ms -> 54.5 ms +- 0.8 ms: 1.04x faster
Significant (t=17.40)
--
### logging_silent ###
Mean +- std dev: 73.1 ns +- 1.2 ns -> 75.9 ns +- 9.0 ns: 1.04x slower
Significant (t=-3.38)
--
### meteor_contest ###
Mean +- std dev: 73.6 ms +- 0.7 ms -> 72.2 ms +- 0.8 ms: 1.02x faster
Significant (t=15.82)
--
### nqueens ###
Mean +- std dev: 61.7 ms +- 0.8 ms -> 63.3 ms +- 0.7 ms: 1.03x slower
Significant (t=-15.58)
--
### pickle_pure_python ###
Mean +- std dev: 248 us +- 5 us -> 242 us +- 4 us: 1.02x faster
Significant (t=10.16)
--
### raytrace ###
Mean +- std dev: 213 ms +- 3 ms -> 207 ms +- 10 ms: 1.03x faster
Significant (t=5.70)
--
### regex_effbot ###
Mean +- std dev: 2.11 ms +- 0.06 ms -> 2.15 ms +- 0.15 ms: 1.02x slower
Significant (t=-3.10)
--
### richards_super ###
Mean +- std dev: 36.9 ms +- 3.1 ms -> 36.1 ms +- 0.6 ms: 1.02x faster
Significant (t=2.83)
--
### scimark_monte_carlo ###
Mean +- std dev: 49.5 ms +- 0.8 ms -> 47.7 ms +- 0.8 ms: 1.04x faster
Significant (t=17.27)
--
### spectral_norm ###
Mean +- std dev: 77.3 ms +- 2.2 ms -> 72.2 ms +- 0.9 ms: 1.07x faster
Significant (t=23.55)
--
### sqlglot_v2_parse ###
Mean +- std dev: 971 us +- 11 us -> 936 us +- 14 us: 1.04x faster
Significant (t=20.56)
--
### sqlglot_v2_transpile ###
Mean +- std dev: 1.20 ms +- 0.01 ms -> 1.16 ms +- 0.02 ms: 1.03x faster
Significant (t=22.28)
--
### subparsers ###
Mean +- std dev: 15.2 ms +- 0.2 ms -> 14.8 ms +- 0.1 ms: 1.03x faster
Significant (t=23.29)
--
### unpack_sequence ###
Mean +- std dev: 25.0 ns +- 0.6 ns -> 26.8 ns +- 0.3 ns: 1.07x slower
Significant (t=-31.46)
--
### unpickle_list ###
Mean +- std dev: 3.52 us +- 0.05 us -> 3.44 us +- 0.06 us: 1.02x faster
Significant (t=10.27)
--
### xml_etree_iterparse ###
Mean +- std dev: 63.8 ms +- 0.7 ms -> 65.1 ms +- 0.4 ms: 1.02x slower
Significant (t=-17.91)
--
### xml_etree_parse ###
Mean +- std dev: 92.7 ms +- 2.1 ms -> 94.9 ms +- 0.7 ms: 1.02x slower
Significant (t=-10.59)
```

</details>

However, I still don't have much confidence in the reliability of these results. Many are unstable. `unpack_sequence`, regressed 7% here, improving 3% when I benchmarked it earlier, and regressing 31% before that (all comparing the same two versions). When I profile it, GC doesn't appear in the `perf report`.